Pre-allocate capacity for maps in json! macro.
Attempt to address issue #810.

This implementation works by expanding the contents of the map twice,
first to increment a capacity variable for each element, then actually
inserting them.

In debug mode, this implementation does not appear to optimize out the incrementations
of the capacity variable, so it is likely not a good solution.